requirements.txt
streamlit>=1.28.0
openai>=1.0.0
Pillow>=9.0.0
pandas>=1.5.0
pdfplumber>=0.7.0
python-dotenv>=1.0.0